(#17) fix: DirectionGap props 역할 분할
