+++ src/pages/auth/SignUp.tsx
@@ -1,11 +1,12 @@
 import { Button } from '@/components/common/button';
 import { Input } from '@/components/common/input';
+import { PasswordInput } from '@/components/common/input/PasswordInput';
 import { AuthWrapper } from '@/layouts/AuthWrapper';
 import { useSignUp } from '@/hooks/useAuth';
 import { useForm } from '@/hooks/useForm';
 import { Text } from '@/components/common/text';
 import { Link } from 'react-router-dom';
-import { ColumnGap } from '@/layouts/ColumnGap';
+import { ColumnCenterGap } from '@/layouts/DirectionGap';
 
 /** 프로필, 이메일 인증 추가 요망 */
 export const SignUpPage = () => {
@@ -21,7 +22,7 @@ export const SignUpPage = () => {
   return (
     <AuthWrapper onSubmit={() => signUpMutate.mutate()}>
       <Text size="heading1">Sign up</Text>
-      <ColumnGap gap="15px" margin="45px 0 40px 0">
+      <ColumnCenterGap gap="15px" margin="45px 0 40px 0">
         <Input
           value={text.name}
           name="name"
@@ -38,10 +39,9 @@ export const SignUpPage = () => {
           onChange={handleOnChange}
           placeholder="이메일을 입력해주세요."
         />
-        <Input
+        <PasswordInput
           value={text.password}
           name="password"
-          type="password"
           label="비밀번호"
           onChange={handleOnChange}
           placeholder="비밀번호를 입력해주세요."
@@ -52,17 +52,16 @@ export const SignUpPage = () => {
           }
           errorMes="비밀번호가 너무 짧거나 특수 문자를 넣어주세요"
         />
-        <Input
+        <PasswordInput
           value={text.passwordCheck}
           name="passwordCheck"
-          type="password"
           label="비밀번호 확인"
           onChange={handleOnChange}
           placeholder="비밀번호를 한 번 더 입력해주세요."
           error={text.passwordCheck !== text.password}
           errorMes="비밀번호가 틀립니다."
         />
-      </ColumnGap>
+      </ColumnCenterGap>
       <Button>회원가입</Button>
       <Text margin="20px 0 0 0" align="center" size="body1" underLine={true}>
         <Link to="/signin">로그인</Link> |{' '}
